[IMP] tests: add dynamic config files
A new mechanism was added on runbot, allowing to specify how to run the
tests from a json file located in the code.

Those file are currently located in odoo/tests/runbot/ but this could
change in the future.

The main motivation is to be able to define how to split the modules
from the odoo code, allowing easier testing and different split per
versions.
This should also help external runbot deployment to use similar config
without needing to create the same config, either by copying the file
of using the config extension mechanism (change the modules to install
as an example)
Finally, some basic config will be added in the code to help to debug
some use cases (enabling the profiler, running basic multibuild,
faketime,...)

More config will be moved from runbot database to the community code in
the future.

Note that this version is only a prototype and may change in the
following weeks.

Part-of: odoo/odoo#234132
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>